Feature: `shouldPostfixNamespaces` flag
This changeset adds a flag to the Soy compiler, called
`shouldPostfixNamespaces`. It defaults to `true` to maintain
current behavior. When `false` is passed via the CLI or as an
option in `SoyJsSrcOptions` or `SoyIncrementalDomSrcOptions`,
the compiler won't postfix paths with, say, `.incrementaldom`.

This allows a user to opt-in, or rather, opt-out, of the default
behavior which allows intermingling templates of different types,
if they do not need to intermingle templates.

Changes so far:
- [x] Add `shouldPostfixNamespaces` option to `SoyJsSrcOptions`
- [x] Add code to `GenIncrementalDomCodeVisitor` to respect the
      new `shouldPostfixNamespaces` flag