Feature: Protos in Soy Python
This changeset extends the compiler with basic support for access
to proto-defined fields. It simply appends normal field accesses
after resolving the object itself from the context `dict`.

I fully expect this change to be dangerous, and it has no
accompanying unit tests yet. But it's better than an error message
if you happen to access a proto field ¯\_(ツ)_/¯.

Changes so far:
- [x] Emit field accesses for protos
- [x] Poorly attempt to make it safer